+++ util/isOutputAsset.ts
@@ -1,0 +1,6 @@
+import { OutputAsset } from "../deps.ts";
+
+// deno-lint-ignore no-explicit-any
+export function isOutputAsset(x: any): x is OutputAsset {
+  return Boolean(x?.isAsset);
+}