+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Directory loading to support dynamic imports and code splitting
+- BREAKING: pluginTerserTransform is now an actual transform plugin, not an output plugin
+- Sourcemapping identifiers are now appended in CLI, not in pluginTerserTransform
 
 ## [0.4.0](https://github.com/denofn/denopack/compare/0.3.2..0.4.0) - 2020-08-11